+++ pybop/optimisers/base_optimiser.py
@@ -353,6 +353,7 @@ class OptimisationResult:
         x: Union[Inputs, np.ndarray] = None,
         final_cost: Optional[float] = None,
         n_iterations: Optional[int] = None,
+        n_evaluations: Optional[int] = None,
         time: Optional[float] = None,
         scipy_result=None,
     ):
@@ -368,6 +369,7 @@ class OptimisationResult:
             else self._calculate_final_cost()
         )
         self.n_iterations = n_iterations
+        self.n_evaluations = n_evaluations
         self.scipy_result = scipy_result
         self.time = time
         if isinstance(self.optim, BaseOptimiser):
@@ -450,6 +452,7 @@ class OptimisationResult:
             f"  Final cost: {self.final_cost}\n"
             f"  Optimisation time: {self.time} seconds\n"
             f"  Number of iterations: {self.n_iterations}\n"
+            f"  Number of evaluations: {self.n_evaluations}\n"
             f"  SciPy result available: {'Yes' if self.scipy_result else 'No'}"
         )
 
@@ -545,6 +548,10 @@ class MultiOptimisationResult:
     @property
     def n_iterations(self):
         return self.best_run().n_iterations
+
+    @property
+    def n_evaluations(self):
+        return self.best_run().n_evaluations
 
     @property
     def scipy_result(self):